SimpleFunctionAndFilterRector
