+++ config/set/twig134.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use Rector\Twig\Rector\SimpleFunctionAndFilterRector;
+use Rector\Twig\Rector\Return_\SimpleFunctionAndFilterRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
