+++ rules/twig/tests/Rector/Return_/SimpleFunctionAndFilterRector/Fixture/fixture.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Twig\Tests\Rector\SimpleFunctionAndFilterRector\Fixture;
+namespace Rector\Twig\Tests\Rector\Return_\SimpleFunctionAndFilterRector\Fixture;
 
 final class SomeExtension extends \Twig_Extension
 {
@@ -16,7 +16,7 @@ final class SomeExtension extends \Twig_Extension
 -----
 <?php
 
-namespace Rector\Twig\Tests\Rector\SimpleFunctionAndFilterRector\Fixture;
+namespace Rector\Twig\Tests\Rector\Return_\SimpleFunctionAndFilterRector\Fixture;
 
 final class SomeExtension extends \Twig_Extension
 {
