+++ rules/twig/tests/Rector/Return_/SimpleFunctionAndFilterRector/Source/TwigSimpleFunction.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Twig\Tests\Rector\Return_\SimpleFunctionAndFilterRector\Source;
+
+final class TwigSimpleFunction
+{
+
+}
